requirements.txt
# Azure OpenAI and Search dependencies
openai>=1.12.0
azure-search-documents>=11.4.0
azure-identity>=1.15.0

# Environment and configuration
python-dotenv>=1.0.0

# Web framework (for API endpoints)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# HTTP requests and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.2
selenium>=4.15.0
webdriver-manager>=4.0.0
readability-lxml>=0.8.1
lxml>=4.9.0

# Logging and utilities
rich>=13.0.0
typing-extensions>=4.8.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# Database dependencies
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0